{{cookiecutter.project_name}}/requirements/base.txt
Django>=2.0.0
{%- if cookiecutter.use_allauth == "y" %}
django-allauth>=0.35.0
{%- endif %}
django-braces>=1.12.0
{%- if cookiecutter.use_celery == "y" %}
django-celery-results>=1.0.1
{%- endif %}
{%- if cookiecutter.use_constance == "y" %}
django-constance[database]>=2.1.0
{%- endif %}
{%- if cookiecutter.use_cors == "y" %}
django-cors-headers>=2.1.0
{%- endif %}
django-environ>=0.4.4
django-extensions>=2.0.0
django-model-utils>=3.1.1
django-silk>=2.0.0
django-solo>=1.1.3
{% if cookiecutter.use_rest == "y" %}
djangorestframework>=3.7.7
djangorestframework-jwt>=1.11.0
{% endif %}
{%- if cookiecutter.use_rest == "y" and cookiecutter.use_allauth == "y" %}
django-rest-auth>=0.9.3
{% endif %}

{%- if cookiecutter.use_celery == "y" %}
celery>=4.1.0
{%- endif %}
{%- if cookiecutter.use_rest == "y" %}
coreapi>=2.3.3
{%- endif %}
{%- if cookiecutter.use_celery == "y" %}
flower>=0.9.2
{%- endif %}
funcy>=1.10.1
marshmallow>=2.15.0
{%- if cookiecutter.use_rabbitmq == "y" %}
pika>=0.11.2
{%- endif %}
Pillow>=5.0.0
psycopg2-binary>=2.7.3.2
psycopg2==2.7.4
Pygments>=2.2.0
{%- if cookiecutter.use_memcached == "y" %}
python-memcached==1.59
{%- endif %}
pytz>=2018.3
raven>=6.5.0
requests>=2.18.4
sqlformatter>=1.3
sqlparse>=0.2.4
